+++ bdc/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Version 1.10.0
 
+* A new `bdc_min_version` configuration item is now _required_. It identifies
+  the minimum version of `bdc` required to parse a particular `build.yaml`.
 * Added `variables` section to `build.yaml`, allowing definition of arbitrary
   variables for substitution. See the sample `build.yaml` for details.
 * `bdc` now automatically generates a top-level version notebook in the DBC